+++ client/src/components/ProfileContainer.tsx
@@ -10,7 +10,6 @@ export const ProfileContainer = () => {
   const params = useParams();
 
   const fetchData = async () => {
-    console.log(params.username);
     const response = await axios.get(
       `http://localhost:3001/user/${params.username}`,
       {
